LEARNIX-97 Moved to Components folder and service folder
