+++ src/app/app.routes.ts
@@ -6,12 +6,14 @@ import {PageNotFoundComponent} from "./Components/page-not-found/page-not-found.
 import {authGuard, authGuardAdmin} from "./auth.guard";
 import {LectionlistComponent} from "./Components/lectionlist/lectionlist.component";
 import {UserprofileComponent} from "./Components/userprofile/userprofile.component";
-import {WorkspacelistComponent} from "./workspacelist/workspacelist.component";
+
 import {AdminConsoleComponent} from "./Components/admin-console/admin-console.component";
 import {PrivacyPolicyComponent} from "./Components/privacy-policy/privacy-policy.component";
 import {TermsOfServiceComponent} from "./Components/terms-of-service/terms-of-service.component";
 import {MainViewComponent} from "./main-view/main-view.component";
-import {WorkspaceViewComponent} from "./workspace-view/workspace-view.component";
+import {WorkspacelistComponent} from "./Components/workspacelist/workspacelist.component";
+import {WorkspaceViewComponent} from "./Components/workspace-view/workspace-view.component";
+
 
 export const routes: Routes = [
   { path: '', component: LandingpageComponent},
